First attempt to categorize all
This text file will be used to reformat the profile config files and
provide a table of contents, making it easier to edit and understand
them.

Also check in new_overlay.ini template which I forgot to add to git; its
organization isn't quite the same and maybe I should unify them.