First version of reformat_config script
This script places printer profile keys in a consistent order and groups
them by logical categories.

I would like to use TOML as an input file so I can have line comments
and things like that; unfortunately because of some inconsistencies in
how PrusaSlicer handles quoting and things, that will be a larger task
that will have to wait.