Rewire how actions are called
Instead of a list of if statements, we now have an ordered dict of
Actions with attributes that the 3dm.py goes through. This means you
should be able to wire in most new actions without altering 3dm.py at
all.

This is a major refactor and will need more thorough testing than I've
done so far.