Standardize capitalization of 3DMake
In order to decide on this, I tried different case combinations in NVDA
and Orca with different synthesizers. Orca and NVDA with espeak
pronounced them all as "three-d make", but with the default Windows
synthesizer option NVDA called 3dmake and 3Dmake "threedmake" which is
undesirable, so I am standardizing on 3DMake. The TOML file names and
things will remain lowercase.