Support edit_in_background
When using a graphical editor, it's annoying to have it the terminal so
you have to close it to use 3dmake again. But for terminal editors, we
can't launch in the background as we need stdio to remain connected.

This was the best solution I could find for that. I would rather
auto-detect whether the child opens a window or something like that, but
it seems brittle and difficult to implement.