Add library downloader
I haven't set this up in the build process yet, but it seems to download
libraries correctly.

The list of libs in library_catalog.toml is pretty barebones right now
because I have to manually add them. I'll probably start with a small
curated list and add more as demand arises.