Add fallback prompt when blocking editor is needed
In case an editor is launched in a workflow where we need the editor to
block unitl we are done (currently just edit-profile-gcode), we want an
editor command in the config that blocks. But in case the user chose one
that doesn't block; we can detect that it closed too quickly and ask the
user to press enter before we continue.